% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\docType{data}
\name{Account}
\alias{Account}
\title{Account}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Account
}
\arguments{
\item{accountID}{accountID}

\item{balance}{accountID}

\item{currency}{currency}
}
\value{
An instance of the R6 class with the predefined fields.
}
\description{
Creating an instance of the R6 class with the predefined fields.
}
\examples{

}
\keyword{datasets}
