% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transaction.R
\docType{data}
\name{Transaction}
\alias{Transaction}
\title{Transaction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Transaction
}
\arguments{
\item{accountID}{accountID}

\item{bic}{bic}

\item{date}{date}

\item{authorization}{authorization}

\item{xRequestId}{xRequestId}

\item{consentId}{consentId}

\item{dateFrom}{dateFrom}

\item{dateTo}{dateTo}
}
\value{
An instance of the R6 class with the predefined fields.
}
\description{
Creating an instance of the R6 class with the predefined fields.
}
\examples{

}
\keyword{datasets}
